+++ src/Tensor.php
@@ -283,6 +283,22 @@ interface Tensor extends ArrayAccess, IteratorAggregate, Countable
     public function clip(float $min, float $max);
 
     /**
+     * Clip the tensor to be lower bounded by a given minimum.
+     *
+     * @param  float  $min
+     * @return mixed
+     */
+    public function clipLower(float $min);
+
+    /**
+     * Clip the tensor to be upper bounded by a given maximum.
+     *
+     * @param  float  $max
+     * @return mixed
+     */
+    public function clipUpper(float $max);
+
+    /**
      * Return the element-wise sign indication.
      * 
      * @return mixed